Fix issues in AuthO router
Every register button redirects to the /signup route, whereas the router signup route is the signin page (as the signup page is the login page, you can see the implementation inside the signup page, so I request to change the name of the page as it is misleading).

So what I have done is change the component of the signup route to be the register page and make a new route that is the login route, which has a login component.